fix bug in OutputCOnfigs; move VarCuts to library
